Refactor the Async Task for getting the city list.
