+++ app/src/main/java/com/example/xweather_v2/MainActivity.java
@@ -3,6 +3,7 @@ package com.example.xweather_v2;
 import android.annotation.SuppressLint;
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
@@ -11,13 +12,22 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 import androidx.viewpager.widget.ViewPager;
 
+import com.example.xweather_v2.bean.CityListBean;
 import com.example.xweather_v2.city_manager.CityManagerActivity;
 import com.example.xweather_v2.db.DatabaseManager;
 import com.example.xweather_v2.today_weather.CityFragmentPagerAdapter;
 import com.example.xweather_v2.today_weather.CityWeatherFragment;
+import com.google.gson.Gson;
+import com.google.gson.reflect.TypeToken;
+import com.label305.asynctask.SimpleAsyncTask;
 
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.zip.GZIPInputStream;
 
 public class MainActivity extends AppCompatActivity implements View.OnClickListener {
 
@@ -30,17 +40,22 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     List<ImageView> imageViewList;
     private CityFragmentPagerAdapter adapter;
 
+    public static List<CityListBean> cityListBeanList;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
+
+        // Use Async Task to get the city list first.
+        new loadCities().execute();
 
         main_img_add_city = findViewById(R.id.main_img_add_city);
         main_img_more = findViewById(R.id.main_img_more);
         main_layout_point = findViewById(R.id.main_layout_point);
         main_vp = findViewById(R.id.main_vp);
 
-//        添加点击事件
+        // Add click events.
         main_img_add_city.setOnClickListener(this);
         main_img_more.setOnClickListener(this);
 
@@ -60,7 +75,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     }
 
     private void initPager() {
-        /* 创建Fragment对象，添加到ViewPager数据源当中*/
+        // Create Fragment objects, add them to the view pager data source.
         for (int i = 0; i < cityList.size(); i++) {
             CityWeatherFragment cityWeatherFragment = new CityWeatherFragment();
             Bundle bundle = new Bundle();
@@ -115,5 +130,53 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
                 break;
         }
         startActivity(intent);
+    }
+
+    //  Get the city list for other activity to use later.
+    private class loadCities extends SimpleAsyncTask<List<CityListBean>> {
+
+        @Override
+        protected List<CityListBean> doInBackgroundSimple() {
+            try {
+                StringBuilder builder = new StringBuilder();
+                InputStream is = getResources().openRawResource(R.raw.city_list);
+                GZIPInputStream gzipInputStream = new GZIPInputStream(is);
+
+                InputStreamReader reader = new InputStreamReader(gzipInputStream);
+                BufferedReader in = new BufferedReader(reader);
+                String readed;
+                while ((readed = in.readLine()) != null)
+                    builder.append(readed);
+
+//                The city_list.json from open weather map is a JSON[] array.
+//                Example:
+//               [
+//                {
+//                    "id": 833,
+//                        "name": "Ḩeşār-e Sefīd",
+//                        "state": "",
+//                        "country": "IR",
+//                        "coord": {
+//                    "lon": 47.159401,
+//                            "lat": 34.330502
+//                },
+//                ]
+
+                cityListBeanList = new Gson().fromJson(builder.toString(), new TypeToken<List<CityListBean>>() {
+                }.getType());
+
+                Log.d("city_list", builder.toString());
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+            return cityListBeanList;
+        }
+
+        @Override
+        protected void onSuccess(List<CityListBean> beanList) {
+            super.onSuccess(beanList);
+
+        }
+
     }
 }
